incorporate review comments and move the old unit tests from block to feature folder
