+++ libs/scripts/delayed.js
@@ -61,8 +61,9 @@ const loadDelayed = ([
     loadPrivacy(getConfig, loadScript);
     loadJarvisChat(getConfig, getMetadata, loadScript, loadStyle);
     if (getMetadata('interlinks') === 'on') {
-      const path = `${getConfig().locale.contentRoot}/keywords.json`;
-      const language = getConfig().locale.lang ?? getConfig().locale.ietf?.split('-')[0];
+      const { locale } = getConfig();
+      const path = `${locale.contentRoot}/keywords.json`;
+      const language = locale.lang ?? locale.ietf?.split('-')[0];
       import('../features/interlinks.js').then((mod) => { mod.default(path, language); resolve(mod); });
     } else {
       resolve(null);
