--- test/blocks/interlinks/mocks/keywords.json
@@ -1,64 +1,0 @@
-{
-  "total": 8,
-  "offset": 0,
-  "limit": 8,
-  "data": [
-    {
-      "Cloud": "DC",
-      "Keyword": "ocean",
-      "URL": "http://oceandecade.org/",
-      "Last Update": "44609",
-      "Notes (SEO Team)": ""
-    },
-    {
-      "Cloud": "DX",
-      "Keyword": "citizens",
-      "URL": "https://www.gartner.com/en/publications/accelerate-digital-for-future-ready-government",
-      "Last Update": "44609",
-      "Notes (SEO Team)": ""
-    },
-    {
-      "Cloud": "CC",
-      "Keyword": "edtech tools",
-      "URL": "https://edex.adobe.com/",
-      "Last Update": "44609",
-      "Notes (SEO Team)": ""
-    },
-    {
-      "Cloud": "PACOM",
-      "Keyword": "work together",
-      "URL": "https://business.adobe.com/products/workfront/main.html",
-      "Last Update": "44609",
-      "Notes (SEO Team)": ""
-    },
-    {
-      "Cloud": "SEO",
-      "Keyword": "creative goals",
-      "URL": "https://www.adobe.com/products/photoshop.html",
-      "Last Update": "44609",
-      "Notes (SEO Team)": ""
-    },
-    {
-      "Cloud": "SEO",
-      "Keyword": "pdf",
-      "URL": "https://www.adobe.com/acrobat/about-adobe-pdf.html",
-      "Last Update": "44606",
-      "Notes (SEO Team)": ""
-    },
-    {
-      "Cloud": "SEO",
-      "Keyword": "a/b testing",
-      "URL": "https://experienceleague.adobe.com/docs/target/using/activities/abtest/test-ab.html?lang=en#types",
-      "Last Update": "44354",
-      "Notes (SEO Team)": ""
-    },
-    {
-      "Cloud": "SEO",
-      "Keyword": "see",
-      "URL": "https://business.adobe.com/",
-      "Last Update": "44706",
-      "Notes (SEO Team)": ""
-    }
-  ],
-  ":type": "sheet"
-}
